requirements.txt
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
huggingface_hub>=0.20.3
requests>=2.26.0
tqdm